[EXAMPLES] * NEW: The WebBrowser example now uses the qt4 webkit component.
[WIKI CGI SCRIPT]
* NEW: Add hungarian to the list of languages.

[GB.DESKTOP]
* BUG: Fix a 64-bit OS related bug when getting a window property.

[GB.QT4]
* NEW: TabStrip[].Children[] now returns a child of a specific TabStrip tab
  from its index.

[GB.QT4.WEBKIT]
* NEW: A new component that has one control, WebView, which is a full
  WebKit browser.



git-svn-id: svn://localhost/gambas/trunk@2150 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.desktop/src/x11.c
@@ -145,7 +145,7 @@ char *X11_get_property(Window wid, Atom prop, Atom *type, int *format, int *pcou
 	
 	*pcount += count;
 	
-	size = *format == 32 ? sizeof(int) : ( *format == 16 ? sizeof(short) : 1 );
+	size = *format == 32 ? sizeof(long) : ( *format == 16 ? sizeof(short) : 1 );
 	
 	GB.FreeString(&_property_value);
 	GB.NewString(&_property_value, (char *)data, count * size);
